checkpoint 1 created mainpage
For this checkpoint, we’ve completed the Landing Page UI, which includes a Home , about , contact , patient login and Doctor login

- log in pages

These are currently presented as visual components within the Landing Page to give a clear idea of the full product scope. Full feature functionality and integration into the main Dashboard will follow in the next phase.